refactor: decompose FinalizeTxUpdater
